chore: rename SpuSocket to StreamSocket
