+++ crates/fluvio/src/spu.rs
@@ -7,16 +7,13 @@ use async_trait::async_trait;
 
 use fluvio_protocol::record::ReplicaKey;
 use fluvio_protocol::api::Request;
-use fluvio_protocol::api::RequestMessage;
 use fluvio_types::SpuId;
 use fluvio_socket::{
-    Versions, VersionedSerialSocket, ClientConfig, MultiplexerSocket, SharedMultiplexerSocket,
-    SocketError, AsyncResponse,
+    VersionedSerialSocket, ClientConfig, MultiplexerSocket, SocketError, AsyncResponse,
 };
+use crate::stream_socket::StreamSocket;
 use crate::FluvioError;
 use crate::sync::MetadataStores;
-
-const DEFAULT_STREAM_QUEUE_SIZE: usize = 10;
 
 /// used for connecting to spu
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
@@ -43,61 +40,11 @@ pub trait SpuDirectory {
         R: Sync + Send;
 }
 
-/// Stream Socket to SPU
-#[derive(Debug)]
-pub struct SpuSocket {
-    config: Arc<ClientConfig>,
-    socket: SharedMultiplexerSocket,
-    versions: Versions,
-}
-
-impl SpuSocket {
-    pub fn new(
-        config: Arc<ClientConfig>,
-        socket: SharedMultiplexerSocket,
-        versions: Versions,
-    ) -> Self {
-        Self {
-            config,
-            socket,
-            versions,
-        }
-    }
-
-    pub async fn create_serial_socket(&mut self) -> VersionedSerialSocket {
-        VersionedSerialSocket::new(
-            self.socket.clone(),
-            self.config.clone(),
-            self.versions.clone(),
-        )
-    }
-
-    pub fn is_stale(&self) -> bool {
-        self.socket.is_stale()
-    }
-
-    pub async fn create_stream_with_version<R: Request>(
-        &mut self,
-        request: R,
-        version: i16,
-    ) -> Result<AsyncResponse<R>, FluvioError> {
-        let mut req_msg = RequestMessage::new_request(request);
-        req_msg.header.set_api_version(version);
-        req_msg
-            .header
-            .set_client_id(self.config.client_id().to_owned());
-        self.socket
-            .create_stream(req_msg, DEFAULT_STREAM_QUEUE_SIZE)
-            .await
-            .map_err(|err| err.into())
-    }
-}
-
 /// connection pool to spu
 pub struct SpuPool {
     config: Arc<ClientConfig>,
     pub(crate) metadata: MetadataStores,
-    spu_clients: Arc<Mutex<HashMap<SpuId, SpuSocket>>>,
+    spu_clients: Arc<Mutex<HashMap<SpuId, StreamSocket>>>,
 }
 
 impl Drop for SpuPool {
@@ -123,7 +70,7 @@ impl SpuPool {
 
     /// create new spu socket
     #[instrument(skip(self))]
-    async fn connect_to_leader(&self, leader: SpuId) -> Result<SpuSocket, FluvioError> {
+    async fn connect_to_leader(&self, leader: SpuId) -> Result<StreamSocket, FluvioError> {
         let spu = self.metadata.spus().look_up_by_id(leader).await?;
 
         let mut client_config = self.config.with_prefix_sni_domain(spu.key());
@@ -141,11 +88,11 @@ impl SpuPool {
         client_config.set_addr(spu_addr);
         let versioned_socket = client_config.connect().await?;
         let (socket, config, versions) = versioned_socket.split();
-        Ok(SpuSocket {
-            socket: MultiplexerSocket::shared(socket),
+        Ok(StreamSocket::new(
             config,
+            MultiplexerSocket::shared(socket),
             versions,
-        })
+        ))
     }
 
     #[instrument(skip(self))]
